Refactoring a bit in musit-service. Better separation of fake and dataporten authentication mechanisms.
