+++ musit-service/src/main/scala/no/uio/musit/security/dataporten/DataportenModule.scala
@@ -17,10 +17,11 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package no.uio.musit.security
+package no.uio.musit.security.dataporten
 
 import com.google.inject.AbstractModule
 import net.codingwell.scalaguice.ScalaModule
+import no.uio.musit.security.{AuthGroupResolver, Authenticator}
 
 class DataportenModule extends AbstractModule with ScalaModule {
