+++ musit-service/src/main/scala/no/uio/musit/security/dataporten/DataportenProvider.scala
@@ -17,7 +17,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package no.uio.musit.security
+package no.uio.musit.security.dataporten
 
 object DataportenProvider {
