+++ service_actor/conf/application.test.conf
@@ -7,7 +7,7 @@ play {
 
   http.errorHandler = "no.uio.musit.service.ErrorHandler"
 
-  modules.enabled += no.uio.musit.security.FakeModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
 }
 
 akka {
