+++ service_actor/test/controllers/UserIntegrationSpec.scala
@@ -20,7 +20,7 @@
 package controllers
 
 import no.uio.musit.security.BearerToken
-import no.uio.musit.security.FakeAuthenticator.fakeAccessTokenPrefix
+import no.uio.musit.security.fake.FakeAuthenticator.fakeAccessTokenPrefix
 import no.uio.musit.test.MusitSpecWithServerPerSuite
 import org.scalatest.time.{Millis, Seconds, Span}
 import play.api.http.Status
