+++ service_auth/conf/application.test.conf
@@ -10,7 +10,7 @@ play {
 
   http.errorHandler = "no.uio.musit.service.ErrorHandler"
 
-  modules.enabled += no.uio.musit.security.FakeModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
 }
 
 akka {
