+++ service_geo_location/conf/application.conf
@@ -7,8 +7,8 @@ play {
 
   http.errorHandler = "no.uio.musit.service.ErrorHandler"
 
-  modules.enabled += no.uio.musit.security.FakeModule
-  //  modules.enabled += no.uio.musit.security.DataportenModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
+  //  modules.enabled += no.uio.musit.security.dataporten.DataportenModule
 
   http.filters= no.uio.musit.service.MusitFilter
 }
