+++ service_storagefacility/conf/application.conf
@@ -24,8 +24,8 @@ play {
   evolutions.db.default.autoApply = true
   evolutions.db.default.autoApply = ${?EVOLUTION_ENABLED}
 
-  modules.enabled += no.uio.musit.security.FakeModule
-  //  modules.enabled += no.uio.musit.security.DataportenModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
+  //  modules.enabled += no.uio.musit.security.dataporten.DataportenModule
 }
 
 akka {
