+++ service_storagefacility/conf/application.test.conf
@@ -10,7 +10,7 @@ play {
   http.errorHandler = "no.uio.musit.service.ErrorHandler"
 
   // Enables the fake MUSIT auth module.
-  modules.enabled += no.uio.musit.security.FakeModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
 }
 
 akka {
