+++ service_thing_aggregate/conf/application.conf
@@ -12,8 +12,8 @@ play {
 
   http.filters= no.uio.musit.service.MusitFilter
 
-  modules.enabled += no.uio.musit.security.FakeModule
-  //  modules.enabled += no.uio.musit.security.DataportenModule
+  modules.enabled += no.uio.musit.security.fake.FakeModule
+  //  modules.enabled += no.uio.musit.security.dataporten.DataportenModule
 }
 
 akka {
