+++ service_thing_aggregate/test/controllers/ObjectAggregationIntegrationSpec.scala
@@ -20,7 +20,8 @@
 package controllers
 
 import no.uio.musit.models.{MuseumNo, ObjectId, SubNo}
-import no.uio.musit.security.{BearerToken, FakeAuthenticator}
+import no.uio.musit.security.BearerToken
+import no.uio.musit.security.fake.FakeAuthenticator
 import no.uio.musit.test.MusitSpecWithServerPerSuite
 import org.scalatest.time.{Millis, Seconds, Span}
 import play.api.libs.json._
